# perf(gatsby): do not call and iterate getAllNodes(File) for each file