--- a/packages/gatsby/src/schema/resolvers.ts
+++ b/packages/gatsby/src/schema/resolvers.ts
@@ -370,7 +370,7 @@ export function fileByPath<TSource, TArgs>(
     )
 
     function queryNodeByPath(
-      relPath: string | string[]
+      relPath: string | Array<string>
     ): Array<Promise<IGatsbyNode>> | Promise<IGatsbyNode> {
       if (Array.isArray(relPath)) {
         const pees = relPath.map(relPath => queryNodeByPath(relPath))