Before it would call ``context.nodeModel.getAllNodes({ type: `File` })`` (already a non-trivial operation on its own) and then filter that down to the one node it was looking for; the one that represents the file. This approach is not cached, super slow, and simply does not scale.

This PR changes that approach to leverage the fast filters, which will generate a O(1) lookup cache on the first call. The rest of the calls will be O(1) lookups.

Tested this on a modification of `benchmarks/gabe-fs-markdown` which adds one image per page (randomly generated before the benchmark).
- At 1000 pages both finish in about 20s
- At 16k pages, from 310s down to 195s
- At 32k pages. from 1057s down to 382s 
- At 64k pages, from 3474s down to 754s
- At 128k pages, from ..3h? down to 1652s

This effect scales as the number of nodes goes up. This should apply to most things that relate to local files, and maybe even beyond that.

# This is a massive win for local file usage

just sayin'